fix(cuda): widen GDN conv1d inner offset to 64-bit
The conv1d forward (linear_attention.cu:110) and backward (:1809) recomputed
the flattened input offset src_t*channels+c in int, overflowing 2^31 for the
Qwen3.6-27B GDN (channels~10240) at src_t>209715 < 262144 — a 256K memory-
corruption wall. The sibling boundary-forward kernel (:143) already used the
wide cast; these two sites were missed by the earlier b3e3f4a22 widening.
Match it: static_cast<unsigned long long>(src_t)*channels+c.

idx/total were already ull; only the recomputed inner index was narrow.

Bench: pending-remote — CUDA does not build on this Mac; measured gate is the
T5 256K OPD ladder (docs/plans/2026-07-29-single-gpu-256k-training-repair.md).

+++ crates/autograd/src/backend_cuda/kernels/linear_attention.cu
@@ -107,7 +107,7 @@ extern "C" __global__ void linear_attention_conv1d_silu_forward_f32_to_bf16(
             }
             continue;
         }
-        sum += x[src_t * channels + c] * weight[c * kernel_size + tap];
+        sum += x[static_cast<unsigned long long>(src_t) * channels + c] * weight[c * kernel_size + tap];
     }
     preact[idx] = sum;
     out_bf16[idx] = la_float_to_bf16(la_silu(sum));
@@ -1806,7 +1806,7 @@ extern "C" __global__ void linear_attention_conv1d_silu_backward_f32(
             }
             continue;
         }
-        int input_idx = src_t * channels + c;
+        unsigned long long input_idx = static_cast<unsigned long long>(src_t) * channels + c;
         float x = input[input_idx];
         float w = weight[c * kernel_size + tap];
         atomicAdd(&grad_input[input_idx], dpre * w);
